extract statement classification helpers
